refactor: reorganize source into feature modules
