+++ environment.md
@@ -1,6 +1,6 @@
 # プロジェクト環境構成
 
-*このファイルは自動的に生成・更新されます*
+_このファイルは自動的に生成・更新されます_
 
 ## 📁 ディレクトリ構造
 
@@ -32,29 +32,66 @@
 ├── src
 │   ├── assets
 │   │   └── img
-│   ├── components
-│   │   ├── Blog
-│   │   ├── Breadcrump
-│   │   ├── Contact
-│   │   ├── Footer
-│   │   ├── Header
-│   │   ├── Hero
-│   │   ├── Profile
-│   │   ├── Section
-│   │   ├── Sns
-│   │   ├── Theme
-│   │   └── ViewToggle.tsx
 │   ├── content
 │   │   ├── blog
 │   │   └── config.ts
-│   ├── layouts
-│   │   └── Layout.astro
+│   ├── features
+│   │   ├── blog
+│   │   │   ├── api
+│   │   │   │   ├── qiita.ts
+│   │   │   │   └── zenn.ts
+│   │   │   └── components
+│   │   │       ├── BlogControls.tsx
+│   │   │       ├── PostCard.tsx
+│   │   │       ├── ShareButtons.astro
+│   │   │       ├── TableOfContents.astro
+│   │   │       └── index.ts
+│   │   ├── contact
+│   │   │   ├── api
+│   │   │   │   └── submitContactForm.ts
+│   │   │   └── components
+│   │   │       ├── ContactForm.tsx
+│   │   │       ├── ContactFormInput.tsx
+│   │   │       ├── ContactFormLabel.tsx
+│   │   │       ├── ContactFormTextarea.tsx
+│   │   │       └── SuspenseContactForm.tsx
+│   │   ├── layout
+│   │   │   ├── components
+│   │   │   │   ├── Breadcrumb
+│   │   │   │   │   └── Breadcrumb.astro
+│   │   │   │   ├── Footer
+│   │   │   │   │   └── Footer.astro
+│   │   │   │   ├── Header
+│   │   │   │   │   └── Header.astro
+│   │   │   │   ├── SocialIcons.astro
+│   │   │   │   └── Theme
+│   │   │   │       └── ThemeToggle.astro
+│   │   │   └── layouts
+│   │   │       └── SiteLayout.astro
+│   │   ├── profile
+│   │   │   ├── components
+│   │   │   │   └── Profile
+│   │   │   │       └── ProfileSection.astro
+│   │   │   └── constants
+│   │   │       └── profileContent.ts
+│   │   └── shared
+│   │       └── components
+│   │           ├── Hero
+│   │           │   └── HeroSection.astro
+│   │           ├── Section
+│   │           │   └── SectionHeading.astro
+│   │           └── ViewToggle.tsx
 │   ├── lib
-│   │   ├── api-clients
 │   │   └── utils
+│   │       └── ogp.ts
 │   ├── pages
 │   │   ├── api
+│   │   │   └── og
+│   │   │       └── [...slug].astro
 │   │   ├── blog
+│   │   │   ├── [...slug].astro
+│   │   │   ├── [...slug].md.ts
+│   │   │   └── index.astro
 │   │   ├── contact.astro
 │   │   ├── index.astro
 │   │   ├── profile.astro
@@ -75,27 +112,28 @@
 ├── tsconfig.json
 └── uno.config.ts
 
-32 directories, 37 files
+33 directories, 40 files
 
 ```
 
 ## 🔑 主要ファイルの役割
 
 ### 設定ファイル
+
 - `astro.config.mjs`: Astroの設定（プラグイン、統合など）
 - `tsconfig.json`: TypeScriptのコンパイラ設定
 
 ### コアコンポーネント
-- `src/layouts/Layout.astro`: 全ページで使用される基本レイアウト
-- `src/components/Header.astro`: サイトヘッダー（ナビゲーション）
-- `src/components/Footer.astro`: サイトフッター
+
+- `src/features/layout/layouts/SiteLayout.astro`: 全ページで使用される基本レイアウト
+- `src/features/layout/components/Header/Header.astro`: サイトヘッダー（ナビゲーション）
+- `src/features/layout/components/Footer/Footer.astro`: サイトフッター
 
 ### ページコンポーネント
 
 ### スタイル
 
 ### ユーティリティ
-
 
 ## 🔄 自動更新の仕組み
