+++ src/features/blog/components/TableOfContents.astro
@@ -1,5 +1,7 @@
 ---
 // Heading interface definition
+import type { JSX } from 'astro/jsx-runtime';
+
 interface Heading {
   depth: number;
   slug: string;
@@ -47,35 +49,24 @@ function buildTocTree(headings: Heading[]): TocNode[] {
   return root.children; // Return children of the root node
 }
 
-// Helper function to render the TOC list recursively
-// Add isNested parameter to control indentation class
-// function renderTocList(nodes: TocNode[], isNested = false): string {
-function renderTocList(nodes: TocNode[]): string {
-  if (!nodes || nodes.length === 0) {
-    return '';
-  }
-
-  const olClass = `list-decimal px-4 leading-7`;
-  let html = `<ol class="${olClass}">`;
-
-  for (const node of nodes) {
-    html += `<li>`;
-    html += `<a href="#${node.slug}" class="text-sm dark:text-white hover:underline transition duration-200">${node.text}</a>`;
-
-    if (node.children.length > 0) {
-      html += renderTocList(node.children);
-    }
-    html += '</li>';
-  }
-
-  html += '</ol>';
-  return html;
-}
+const renderTocItems = (nodes: TocNode[]): JSX.Element[] =>
+  nodes.map((node) => (
+    <li>
+      <a
+        href={`#${node.slug}`}
+        class="text-sm dark:text-white hover:underline transition duration-200"
+      >
+        {node.text}
+      </a>
+      {node.children.length > 0 && (
+        <ol class="list-decimal px-4 leading-7">{renderTocItems(node.children)}</ol>
+      )}
+    </li>
+  ));
 
 // Filter headings to include only h2 and h3, then build the tree
 const filteredHeadings = headings.filter((h) => h.depth >= 2 && h.depth <= 3);
 const tocTree = buildTocTree(filteredHeadings);
-const tocHtml = renderTocList(tocTree); // Initial call, isNested defaults to false
 
 const tocStyles = {
   container: 'mb-8',
@@ -92,7 +83,9 @@ const tocStyles = {
       <span class={tocStyles.arrow}>&#9660;</span>
       目次
     </summary>
-    <nav class={tocStyles.nav} set:html={tocHtml} />
+    <nav class={tocStyles.nav}>
+      {tocTree.length > 0 && <ol class="list-decimal px-4 leading-7">{renderTocItems(tocTree)}</ol>}
+    </nav>
   </details>
 </div>
