+++ src/features/layout/components/Footer/Footer.astro
@@ -1,5 +1,5 @@
 ---
-import SocialIcons from '../Sns/SocialIcons.astro';
+import SocialIcons from '../SocialIcons.astro';
 
 const currentYear = new Date().getFullYear();
