+++ src/features/layout/components/Header/Header.astro
@@ -1,6 +1,6 @@
 ---
 import ThemeToggle from '../Theme/ThemeToggle.astro';
-import SocialIcons from '../Sns/SocialIcons.astro';
+import SocialIcons from '../SocialIcons.astro';
 
 const { pathname } = Astro.url;
