+++ src/features/layout/layouts/SiteLayout.astro
@@ -1,7 +1,7 @@
 ---
 import Header from '../components/Header/Header.astro';
 import Footer from '../components/Footer/Footer.astro';
-import Breadcrumb from '../components/Breadcrump/Breadcrumb.astro';
+import Breadcrumb from '../components/Breadcrumb/Breadcrumb.astro';
 import '@fontsource/zen-kaku-gothic-new/japanese-400.css'; // 日本語サブセット（標準ウェイト）
 import '@fontsource/zen-kaku-gothic-new/japanese-700.css'; // 日本語サブセット（太字）
