+++ src/features/profile/components/Profile/ProfileSection.astro
@@ -1,5 +1,5 @@
 ---
-import { heading } from '../../styles/unoVariants';
+import { heading } from '@/styles/unoVariants';
 
 interface Props {
   title: string;
