+++ src/features/shared/components/Hero/HeroSection.astro
