+++ src/pages/contact.astro
@@ -1,7 +1,7 @@
 ---
-import Layout from '../layouts/Layout.astro';
-import HeroSection from '../components/Hero/HeroSection.astro';
-import SuspenseContactForm from '../components/Contact/SuspenseContactForm';
+import Layout from '@/features/layout/layouts/SiteLayout.astro';
+import HeroSection from '@/features/shared/components/Hero/HeroSection.astro';
+import SuspenseContactForm from '@/features/contact/components/SuspenseContactForm';
 ---
 
 <Layout title="お問い合わせ | SAKUSPACE">
