+++ src/pages/index.astro
@@ -1,11 +1,11 @@
 ---
 import { getCollection } from 'astro:content';
-import Layout from '../layouts/Layout.astro';
-import HeroSection from '../components/Hero/HeroSection.astro';
-import { PostCard } from '../components/Blog/PostCard';
-import SectionHeading from '../components/Section/SectionHeading.astro';
-import { getQiitaPosts } from '../lib/api-clients/qiita';
-import { getZennPosts } from '../lib/api-clients/zenn';
+import Layout from '@/features/layout/layouts/SiteLayout.astro';
+import HeroSection from '@/features/shared/components/Hero/HeroSection.astro';
+import { PostCard } from '@/features/blog/components/PostCard';
+import SectionHeading from '@/features/shared/components/Section/SectionHeading.astro';
+import { getQiitaPosts } from '@/features/blog/api/qiita';
+import { getZennPosts } from '@/features/blog/api/zenn';
 
 const posts = await getCollection('blog');
 const qiitaPosts = await getQiitaPosts('ngtnysk');
